refactor: move SearchDecisionRequirementsTest tests to DecisionQueryTest
